+++ lib/identity_cache/configuration_dsl.rb
@@ -217,6 +217,8 @@ module IdentityCache
         options[:records_variable_name]   = "cached_#{association}"
         options[:prepopulate_method_name] = "prepopulate_fetched_#{association}"
 
+        inverse_name = options[:association_reflection].inverse_of.try(:name)
+
         self.class_eval(<<-CODE, __FILE__, __LINE__ + 1)
           attr_reader :#{options[:ids_variable_name]}
 
@@ -227,7 +229,12 @@ module IdentityCache
           def #{options[:cached_accessor_name]}
             association_klass = association(:#{association}).klass
             if association_klass.should_use_cache? && !#{association}.loaded?
-              @#{options[:records_variable_name]} ||= #{options[:association_reflection].class_name}.fetch_multi(#{options[:cached_ids_name]})
+              @#{options[:records_variable_name]} ||= begin
+                #{options[:association_reflection].class_name}.fetch_multi(#{options[:cached_ids_name]}).map do |record|
+                  record.send('#{inverse_name}=', self) if #{inverse_name.present?}
+                  record
+                end
+              end
             else
               #{association}.to_a
             end